Add files via upload
this file contains the preprocessing steps used to create the variables that went into the model. Three types of variables were created which combined the data from all 3 seasons. The first type of variable was summing the stat selected for all three seasons. Second, some stats had a weighted average calculated for all three years. Finally, the third type was calculating the range of the stat for all three seasons.