Add files via upload
the pdf file shows the results from running the Rmd file. It's quite ugly code, and there are probably not enough comments to explain what's going on. However, it is clear from looking at the results that the model is converging (this is using the normalized response and normal likelihood). Also, from the results, specifically the residual plots, the model is biased and having a hard time modeling the larger AAV values. I used AAV as the response but totalAmount would probably go through the same procedure.